requirements.txt
# =============================================================================
# GABI - Requirements de Produção
# Gerado a partir de pyproject.toml
# =============================================================================

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
sqlalchemy[asyncio]==2.0.28
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.4.2

# Task Queue
celery==5.3.6
redis==5.0.1
flower==2.0.1

# Elasticsearch
elasticsearch[async]==8.11.0

# HTTP Client
httpx==0.27.0
httpx-socks==0.9.0

# Data Validation
pydantic==2.6.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Authentication
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
python-multipart==0.0.9

# Crawling & Parsing
playwright==1.41.2
beautifulsoup4==4.12.3
lxml==5.1.0
html5lib==1.1

# PDF Processing
pdfplumber==0.11.0
pymupdf==1.24.10
pypdf==4.0.2

# Data Processing
pandas==2.2.1
polars==0.20.10
pyarrow==15.0.0
openpyxl==3.1.2
python-magic==0.4.27

# Utilities
python-dotenv==1.0.1
typer==0.9.0
rich==13.7.0
structlog==24.1.0
orjson==3.9.15
ujson==5.9.0
tenacity==8.2.3
backoff==2.2.1

# Observability
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-prometheus==0.43b0

# Security
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
cryptography==42.0.5
itsdangerous==2.1.2

# Text Processing
nltk==3.8.1
spacy==3.7.4
langdetect==1.0.9
unidecode==1.3.8

# YAML
PyYAML==6.0.1
ruamel.yaml==0.18.6

# Async Utilities
aiocache==0.12.2
aiofiles==23.2.1
aiosqlite==0.20.0
asyncio-mqtt==0.16.1
